trail1/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
motor
bcrypt
PyJWT